warenkorb verbessert

+++ system/warenkorb/warenkorbaktionen.php
@@ -1,9 +1,10 @@
 <?php
-// warenkorb klasse starten
-include 'warenkorb.php';
-$warenkorb = new warenkorb;
+include 'warenkorb.php';  // warenkorb klasse starten
+$warenkorb = new warenkorb;  //erzeugt neues Objekt "Warebnkorb"
 
-include '../account/config.php';
+include '../../config.php';
+
+
 if(isset($_GET['action']) && !empty($_GET['action'])) {
     if ($_GET['action'] == 'update_warenkorbartikel' && !empty($_GET['id'])) {
         $artikel_daten = array(
@@ -21,24 +22,27 @@ if(isset($_GET['action']) && !empty($_GET['action'])) {
     }
 }
 
-
-
-
 if(isset($_POST['warenkorb'])){
-    $artikel_id = $_POST['artikel_id'];
+    $artikel_id = $_POST['artikel_id'];  //für artikel_id wird die jeweilige $row["id"] des donuts übergeben
     //Produktdetails abfragen
-    $query = $db->query("SELECT * FROM artikel WHERE id = ".$artikel_id); //alle Infos wo Artikel die ID hat
-    $row = $query->fetch_assoc();
-    $artikel_daten = array(         //Infos mit row rauslesen
-        'id' => $row['id'],
-        'donutname' => $row['donutname'],
-        'preis' => $row['preis'],
-        'menge' => $_POST['mengenangabe'],       //festlegen menge --> ändern
-        'ean'  => $row['ean'],
-        'beschreibung' => $row['beschreibung'],
-        'ende' => $row['ende']
-    );
-
+    $stmt = $db->prepare("SELECT * FROM artikel WHERE id = ".$artikel_id); //alle Infos des Artikels mit der jeweiligen id rauslesen
+    if(!$stmt->execute()) {
+        echo "Datenbank-Fehler ";
+        $arr = $stmt->errorInfo();
+        print_r($arr);
+        die();
+    }
+    while ($row = $stmt->fetch(PDO::FETCH_ASSOC)) {
+        $artikel_daten = array(         //Infos mit row rauslesen
+            'id' => $row['id'],
+            'donutname' => $row['donutname'],
+            'preis' => $row['preis'],
+            'menge' => $_POST['mengenangabe'],       //festlegen menge --> ändern
+            'ean' => $row['ean'],
+            'beschreibung' => $row['beschreibung'],
+            'ende' => $row['ende']
+        );
+    }
     $artikel_einfuegen = $warenkorb->einfuegen($artikel_daten);
     $weiterleiten = $artikel_einfuegen?'../../index.php?page=warenkorbansicht':'../../index.php?page=alledonuts';
     header("Location: ".$weiterleiten);
@@ -47,15 +51,19 @@ if(isset($_POST['warenkorb'])){
 }
 
 
-
-
 if(isset($_POST['bestellung']) && $warenkorb->artikel_gesamt() > 0 && !empty($_SESSION['nutzer']['id'])){  //bestellung aufgeben, wenn warenkorb nicht leer ist und mit nutzer ide
     $bezahlmethode=($_POST['zahlungsinfo']);
         // bestelldaten in DB eintragen
-    $bestellung_einfuegen = $db->query("INSERT INTO bestellungen (benutzer_id, bezahlmethode, endpreis, erstellt, bearbeitet) VALUES ('" . $_SESSION['nutzer']['id'] . "', '" . $bezahlmethode . "', '" . $warenkorb->gesamt() . "', '" . date("Y-m-d H:i:s") . "', '" . date("Y-m-d H:i:s") . "')");
+    $bestellung_einfuegen = $db->prepare("INSERT INTO bestellungen (benutzer_id, bezahlmethode, endpreis, erstellt, bearbeitet) VALUES ('" . $_SESSION['nutzer']['id'] . "', '" . $bezahlmethode . "', '" . $warenkorb->gesamt() . "', '" . date("Y-m-d H:i:s") . "', '" . date("Y-m-d H:i:s") . "')");
+    if(!$bestellung_einfuegen->execute()) {
+        echo "Datenbank-Fehler ";
+        $arr = $bestellung_einfuegen->errorInfo();
+        print_r($arr);
+        die();
+    }
     // wenn bestelldaten eingefügt wurden, artikel daten einfügen
     if ($bestellung_einfuegen) {
-        $bestellungen_id = $db->insert_id;
+        $bestellungen_id = $db->lastInsertId();
         $sql = '';
 
         $warenkorb_artikel = $warenkorb->inhalte();
@@ -63,7 +71,13 @@ if(isset($_POST['bestellung']) && $warenkorb->artikel_gesamt() > 0 && !empty($_S
             $sql .= "INSERT INTO bestellte_artikel (id, bestellungen_id, artikel_id, menge) VALUES ('', '" . $bestellungen_id . "', '" . $artikel['id'] . "', '" . $artikel['menge'] . "');";
         }
         // bestellte artikel in datenbank einfügen
-        $bestellte_artikel_einfuegen = $db->multi_query($sql);
+        $bestellte_artikel_einfuegen = $db->prepare($sql);
+        if(!$bestellte_artikel_einfuegen->execute()) {
+            echo "Datenbank-Fehler ";
+            $arr = $bestellte_artikel_einfuegen->errorInfo();
+            print_r($arr);
+            die();
+        }
 
         if ($bestellte_artikel_einfuegen) {
             $warenkorb->destroy();
